mfdp/requirements.txt
fastapi>=0.68.0,<0.70.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.7.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
python-telegram-bot
gunicorn>=20.1.0,<21.0.0
requests>=2.26.0,<3.0.0
pandas
xlrd
setuptools
pika>=1.3.2,<2.0.0
celery>=5.3.6,<6.0.0
redis>=4.5.4,<5.0.0
rabbitmq
httpx